requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
pillow
keras
tensorflow
numpy